+++ CHANGELOG.md
@@ -7,4 +7,7 @@ All notable changes to this project will be documented in this file.
 - Laravel 7 support by [@divine](https://github.com/divine).
 
 ### Changed
-- Updated versions of all dependencies by [@divine](https://github.com/divine)
+- Updated versions of all dependencies by [@divine](https://github.com/divine).
+
+### Removed
+- shouldUseCollections function by [@divine](https://github.com/divine).